# gui/fastapi/doc/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenTREP Travel Search",
        "description": "REST API for the [OpenTREP](https://github.com/trep/opentrep) open-source travel search engine.\n\nOpenTREP uses Xapian full-text search and SQLite (or MySQL/PostgreSQL) to index and query the [OpenTravelData (OPTD)](https://github.com/opentraveldata/opentraveldata) Points of Reference (POR) dataset, which covers all IATA-registered airports, cities, and heliports worldwide.\n\n**Source**: https://github.com/trep/opentrep  \n**OPTD data**: https://github.com/opentraveldata/opentraveldata",
        "contact": {
            "name": "OpenTREP project",
            "url": "https://github.com/trep/opentrep",
            "email": "denis.arnaud_opentrep@m4x.org"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.7.18"
    },
    "paths": {
        "/api/search": {
            "get": {
                "summary": "Search Points of Reference",
                "description": "Search the OPTD POR dataset using OpenTREP.\n\nThe query can be free-text (city name, airport name) or a sequence of IATA codes. When multiple locations are resolved, the response includes the great-circle distance between consecutive pairs and the total route distance.",
                "operationId": "search_api_search_get",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 200,
                            "description": "Search query: free-text or IATA code(s)",
                            "examples": [
                                "nce",
                                "cdg jfk lax",
                                "Tokyo London"
                            ],
                            "title": "Q"
                        },
                        "description": "Search query: free-text or IATA code(s)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SearchResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/random": {
            "get": {
                "summary": "Random airports / itinerary",
                "description": "Draw *n* random Points of Reference from the OPTD dataset and return the same structure as `/api/search`. Useful for demos and testing.\n\n- `n=1` \u2192 single random airport  \n- `n=3` \u2192 random three-leg itinerary",
                "operationId": "random_airports_api_random_get",
                "parameters": [
                    {
                        "name": "n",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 10,
                            "minimum": 1,
                            "description": "Number of random locations to draw",
                            "default": 1,
                            "title": "N"
                        },
                        "description": "Number of random locations to draw"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SearchResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SearchResponse": {
                "properties": {
                    "query": {
                        "type": "string",
                        "title": "Query",
                        "description": "Query string as received",
                        "examples": [
                            "nce cdg jfk"
                        ]
                    },
                    "locations": {
                        "items": {
                            "$ref": "#/components/schemas/Location"
                        },
                        "type": "array",
                        "title": "Locations",
                        "description": "Ordered list of matched Points of Reference"
                    },
                    "total_km": {
                        "type": "integer",
                        "title": "Total Km",
                        "description": "Total great-circle distance along the full route (km)",
                        "default": 0,
                        "examples": [
                            9808
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "query",
                    "locations"
                ],
                "title": "SearchResponse"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}